requirements.txt
emojis==0.6.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyperclip==1.8.2
PyQt5==5.15.7
Pillow
pilmoji
